fix: update stackage resolver to lts-22.11 for build compatibility
Avoided not being registered in stackage due to the following error.

> ```
> - aws-lambda-haskell-runtime < 0 # tried aws-lambda-haskell-runtime-4.1.2, but its *library* requires the disabled package: safe-exceptions-checked
> ```
>
> <https://github.com/commercialhaskell/stackage/blob/fb2a50c986c1dab18c5aeef5bf1e12eadcbb37f3/build-constraints.yaml#L6267C1-L6267C156>

The [safe-exceptions-checked](https://hackage.haskell.org/package/safe-exceptions-checked) library was deprecated,
causing it not to build.
It has an old list of dependencies,
so I tried adding it to `extra-deps` and found that transformers, etc.
require a large number of out-of-stackage-managed dependencies,
causing it not to build.
I could not find any other good-looking check-exception library.
Having no choice I decided to remove the check exception mechanism itself.

I believe that it should not contain any destructive changes.
The removal of the check exception feature changes the type signature,
but does not change the behavior,
so it is not broken unintentionally.
It is also sufficient for the user to remove the type signatures if they have been written.